separated weapons from units
