+++ CMakeLists.txt
@@ -34,7 +34,7 @@ set(CORE gameManager.cpp game.cpp fxManager.cpp defConfigs.cpp ${CONSOLE} ${CONT
 set(PROJECTILES projectile.cpp missile.cpp cruiseMissile.cpp shell.cpp)
 set(VEHICLES vehicle.cpp submarine.cpp engineer.cpp resourceRover.cpp)
 set(STRUCTURES structure.cpp factory.cpp pointDefense.cpp extractor.cpp researchStruct.cpp iceSheet.cpp)
-set(UNITS gameObjectFactory.cpp unit.cpp ${STRUCTURES} ${VEHICLES})
+set(UNITS gameObjectFactory.cpp unit.cpp weapon.cpp ${STRUCTURES} ${VEHICLES})
 set(CONTENT player.cpp trader.cpp pathfinder.cpp map.cpp gameObject.cpp gameObjectFrame.h resourceDeposit.cpp ${UNITS} ${PROJECTILES})
 set(GAME_SRC ${STATES} ${GUI} ${CORE} ${CONTENT} ${UTIL})
