+++ unit.h
@@ -27,6 +27,7 @@ namespace vb01{
 
 namespace battleship{
     class Player;
+	class Weapon;
 	class Unit;
 	class Vehicle;
 	class Structure;
@@ -75,6 +76,7 @@ namespace battleship{
 	   	CARRIER,
 	   	SUBMARINE,
 		ICEBREAKER,
+		FREEZER,
 	   	LAND_FACTORY,
 	   	NAVAL_FACTORY,
 	   	TRADE_CENTER,
@@ -88,41 +90,6 @@ namespace battleship{
     
     class Unit : public GameObject{
     public:
-		class Weapon{
-			public:
-				Weapon(Unit*, sol::table, int);
-				~Weapon();
-				virtual void update();
-				virtual void fire(Order);
-				void trackTarget(vb01::Vector3);
-				inline int getProjectileId(){return projId;}
-				inline int getRateOfFire(){return rateOfFire;}
-				inline int getDamage(){return damage;}
-				inline int getMinRange(){return minRange;}
-				inline int getMaxRange(){return maxRange;}
-				inline Unit* getUnit(){return unit;}
-				inline Order::TYPE getOrderType(){return orderType;}
-			private:
-				Unit *unit = nullptr;
-				Order::TYPE orderType;
-				int id, projId = -1, rateOfFire, damage = 0;
-				float minRange = 0, maxRange, rotSpeed, maxFireAngle;
-				vb01::s64 lastFireTime = 0;
-				FxManager::Fx *fireFx = nullptr;
-				vb01::Quaternion projRot;
-				vb01::Vector3 projPos;
-				float horConstraint = 1.57, vertConstraint = 1.57;
-				vb01::Node *horNode = nullptr, *vertNode = nullptr, *projPar = nullptr;
-				static std::string LASER_FLAG;
-
-				vb01::Vector3 calcOrientVec(int);
-				void alignNode(vb01::Vector3, vb01::Node*, bool);
-				void initProjectileData(sol::table);
-				vb01::Node* initNode(sol::table, bool);
-				FxManager::Fx* initFx(sol::table, std::string, bool);
-				void useFx(FxManager::Fx*, vb01::Vector3, bool);
-				inline bool canFire(){return vb01::getTime() - lastFireTime > rateOfFire;}
-		};
 
 		struct GarrisonSlot{
 			Vehicle *vehicle = nullptr;
@@ -160,7 +127,7 @@ namespace battleship{
         inline float getLineOfSight() {return lineOfSight;}
         inline UnitType getType() {return type;}
         inline UnitClass getUnitClass() {return unitClass;}
-        inline void takeDamage(int damage) {health -= damage;}
+        inline void takeDamage(int damage) {health -= damage * (1 + (freezeDmgFactor - 1) * freezeStatus * .01f);}
         inline int getPlayerId() {return playerId;}
 		inline int getHealth(){return health;}
 		inline int getDeathHp(){return DEATH_HP;}
@@ -171,6 +138,9 @@ namespace battleship{
 		inline std::string getGuiScreen(){return guiScreen;}
 		inline BuildableUnit getBuildableUnit(int i){return buildableUnits[i];}
 		inline vb01::Node* getLosLightNode(){return losLightNode;}
+		inline void setFreezeStatus(int fs){this->freezeStatus = std::clamp(fs, 0, 100);}
+		inline int getFreezeStatus(){return freezeStatus;}
+		inline bool isEnabled(){return enabled;}
     private:
 		void renderOrderLine(bool);
         void updateScreenCoordinates();
@@ -183,13 +153,13 @@ namespace battleship{
         sf::SoundBuffer *selectionSfxBuffer;
         sf::Sound *selectionSfx = nullptr;
 		vb01::Node *hpBackgroundNode = nullptr, *hpForegroundNode = nullptr, *losLightNode = nullptr;
-		bool vehicle;
+		bool vehicle, enabled = true;
     protected:
         UnitClass unitClass;
         UnitType type;
         std::vector<Order> orders;
 		std::string guiScreen = "";
-        int health = 0, maxHealth, playerId, lenHpBar = 200;
+        int health = 0, maxHealth, playerId, lenHpBar = 200, freezeStatus = 0, freezeDmgFactor = 10;
 		vb01::s64 orderLineDispTime = 0, lastFireTime = 0;
         float lineOfSight;
 		std::vector<Armor> armorTypes;
@@ -219,6 +189,7 @@ namespace battleship{
         virtual void launch(Order);
 		virtual void handleResources(Order){}
 		virtual void hack(Order){}
+		virtual void freeze(Order){}
 		float calculateRotation(vb01::Vector3, float, float);
 		void removeBar(vb01::Node*);
 		vb01::Node* createBar(vb01::Vector2, vb01::Vector2, vb01::Vector4);
